pyproject.toml
[project]
name = "ai-email-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.193.0",
    "google-auth-oauthlib>=1.3.1",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.6",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
